Added report Download/Upload

+++ application/views/admin/templates.php
@@ -1,6 +1,8 @@
 <script type="text/javascript">
     function getNights( prefix ) {
-        var url = prefix + document.getElementById("nights");
+        var e = document.getElementById("nights");
+        var strNights = e.options[e.selectedIndex].value;
+        var url = prefix + strNights;
         window.location.href = url;
     }
 </script>
@@ -12,17 +14,19 @@
         <div class="row-fluid">
           <div class="span4">
              <h2>Templates</h2>
-             <select id='nights'>
-              <option value='1'>1</option>
-              <option value='2'>2</option>
-              <option value='3'>3</option>
-              <option value='4'>4</option>
-             </select>Night(s)
              <div class="pull-right">
                 <button type="button" class="btn btn-default">
-                    <a href="<?= URL::base_uri() . 'admin/templates/add' ?>"><span class="glyphicon glyphicon-plus"></span> <?= $this->lang['add']; ?></a>
+                    <a href="<?= URL::base_uri() . 'admin/templates/x/x/add' ?>"><span class="glyphicon glyphicon-plus"></span> <?= $this->lang['add']; ?></a>
                 </button>
              </div>
+             <div class="input-group">
+                <select id='nights' class="form-control">
+                 <option value='1'>1 Night</option>
+                 <option value='2'>2 Nights</option>
+                 <option value='3'>3 Nights</option>
+                 <option value='4'>4 Nights</option>
+                </select>
+             </div>   
              <table class="table table-hover">
               <?php if ($this->templates == false): ?>
               <p><?= $this->lang['notemplates']; ?></p>
@@ -41,11 +45,13 @@
                   <?php
                     if (is_array($data)){
                         foreach ($data as $part){
-                            $imgu = URL::base_uri() . 'img/flags/';
-                            $url = URL::base_uri() . 'admin/templates';
-                            echo "&nbsp;&nbsp;&nbsp;&nbsp;<img src='$imgu$part.png' alt='$part' />&nbsp;&nbsp;";
-                            echo "<a href=\"javascript:getNights('$url/$list[0]/$part/download/')\">Download</a>/";
-                            echo "<a href=\"javascript:getNights('$url/$list[0]/$part/replace/')\">Replace</a>";
+                            if (is_dir(preg_replace("/\\.docx$/", "", "/var/www/scripts/templates/$list[0]/$part"))){
+                                $imgu = URL::base_uri() . 'img/flags/';
+                                $url = URL::base_uri() . 'admin/templates';
+                                echo "&nbsp;&nbsp;&nbsp;&nbsp;<img src='$imgu$part.png' alt='$part' />&nbsp;&nbsp;";
+                                echo "<a href=\"javascript:getNights('$url/$list[0]/$part/download/')\">Download</a>/";
+                                echo "<a href=\"javascript:getNights('$url/$list[0]/$part/replace/')\">Replace</a>";
+                            }
                         }
                     } else {
                         echo $data;
@@ -53,6 +59,9 @@
                   ?>
                   </td>
                   <?php endforeach; ?>
+                  <td>
+                      <a href="<?= URL::base_uri(); ?>admin/templates/<?= $list[0]; ?>/x/delete" onClick="return confirm('Are you sure?')"><i class="glyphicon glyphicon-remove"></i><?= $this->lang['delete']; ?></a>
+                  </td>
                 </tr>
                 <?php endforeach; ?>
               </tbody>
